Refactor well known extensions using factory pattern
