+++ src/Tools/dotnet-monitor/ServiceCollectionExtensions.cs
@@ -22,7 +22,6 @@ using Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Options.Triggers.Event
 using Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Triggers;
 using Microsoft.Diagnostics.Tools.Monitor.Egress;
 using Microsoft.Diagnostics.Tools.Monitor.Egress.Configuration;
-using Microsoft.Diagnostics.Tools.Monitor.Egress.Extension;
 using Microsoft.Diagnostics.Tools.Monitor.Egress.FileSystem;
 using Microsoft.Diagnostics.Tools.Monitor.Exceptions;
 using Microsoft.Diagnostics.Tools.Monitor.Extensibility;
@@ -233,7 +232,9 @@ namespace Microsoft.Diagnostics.Tools.Monitor
         public static IServiceCollection ConfigureExtensions(this IServiceCollection services)
         {
             services.AddSingleton<ExtensionDiscoverer>();
-            services.AddSingleton<ExtensionRepository, WellKnownEgressExtensionRepository>();
+            services.AddSingleton<ExtensionRepository, WellKnownExtensionRepository>();
+            // Well-known extensions
+            services.AddSingleton<IWellKnownExtensionFactory, FileSystemEgressExtensionFactory>();
             return services;
         }
